backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.19
httpx==0.28.1
yake==0.4.8
reportlab==4.2.5
pillow==11.0.0
python-dotenv==1.0.1
trafilatura==1.12.2
